Renamed code folder to "backend" (requirement)
Also added @Configuration parameter to WebSecurityConfig

Some .css linking fixes
