+++ FASE 2/helloword-vscode/src/main/java/es/codeurjc/backend/Application.java
@@ -1,4 +1,4 @@
-package es.codeurjc.NexusApplication;
+package es.codeurjc.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
