+++ FASE 2/helloword-vscode/src/main/java/es/codeurjc/backend/controller/FeedController.java
@@ -1,11 +1,11 @@
-package es.codeurjc.NexusApplication.controller;
+package es.codeurjc.backend.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import es.codeurjc.NexusApplication.service.TweetService;
+import es.codeurjc.backend.service.TweetService;
 
 @Controller
 public class FeedController {
