+++ FASE 2/helloword-vscode/src/main/java/es/codeurjc/backend/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package es.codeurjc.NexusApplication.repository;
+package es.codeurjc.backend.repository;
 
 import java.util.Optional;
 
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
-import es.codeurjc.NexusApplication.model.User;
+import es.codeurjc.backend.model.User;
 
 @Component
 public interface UserRepository extends JpaRepository<User, Long>{
