+++ FASE 2/helloword-vscode/src/main/java/es/codeurjc/backend/service/TweetService.java
@@ -1,4 +1,4 @@
-package es.codeurjc.NexusApplication.service;
+package es.codeurjc.backend.service;
 
 import org.springframework.stereotype.Component;
